requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.12
pydantic==2.10.3
python-dotenv==1.0.1
pyjwt==2.9.0

